etl/requirements.txt
prefect
polars
psycopg2-binary
numpy
pandas
sentence-transformers
chromadb
python-dotenv
